added floats

+++ 19_calculator_bonus/calculator.py
@@ -11,7 +11,7 @@ with open('calculator_messages.json', 'r') as file:
 
 def invalid_number(number_str):
     try:
-        int(number_str)
+        float(number_str)
     except ValueError:
         return True
 
@@ -60,13 +60,13 @@ while True:
 
     match operation:
         case "1":
-            output = int(number1) + int(number2)
+            output = float(number1) + float(number2)
         case "2":
-            output = int(number1) - int(number2)
+            output = float(number1) - float(number2)
         case "3":
-            output = int(number1) * int(number2)
+            output = float(number1) * float(number2)
         case "4":
-            output = int(number1) / int(number2)
+            output = float(number1) / float(number2)
 
     print(f"{MESSAGES[language_code]['result']} {output}")
